Pass Miri with the implementation
Hui, that is a borrow checking conundrum but we figured it out. The
complex part is that the coroutine can invalidate all forms of pointers
into the IoFuture without communicating with the waker. Thus we must
ensure firstly that the provenance of the access from Waker into the
future is always re-validated and secondly that we do not uniquely
access the future on the exit path (where such a re-validation may not
occur).